chore(tools): remove package next-standard-version
This removes dependency on next-standard-version which has vulnerable
transitive deps and is not actively maintained.

The release script has been updated to use standard-version --dry-run
to accommodate user input to confirm/abort releases. It skips unneeded
automation options for the dry run to draw focus on pending version
change and make it clear to the executor what the new version will be.

Documentation has been updated.

closes #76